requirements.txt
Django==4.1.2
django-cors-headers==3.13.0
django-cuser==2017.3.16
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
Pillow==9.2.0
psycopg2-binary==2.9.3
python-decouple==3.6
gunicorn==20.1.0
django-storages==1.13.2
